don't depend on git to identify files in this gem
This popped up from Kali Linux users using Metasploit who noticed a random git
error message with the latest version of this gem. We traced it back to git
being used unconditionally by this gem in the gemspec file. This changes to
prefer rake's file glob method instead, which removes the dependency on the
.git directory existing.