Address Ruby 2.8 #untaint deprecation warnings
Prior to this commit we were seeing errors when calling the deprecated
Object#untaint method on Ruby 2.8. The method is a no-op on Ruby 2.8.

```rb
dnsruby/config.rb:325: warning: Object#untaint is deprecated and will be removed in Ruby 3.2
```

We probably could remove untainting for older Ruby versions as well, but
it felt safer to preserve the current behavior in case someone is still
relying on it.

To see this change in action, we added ruby-head to the travis matrix,
but do not require those tests to pass for the build as a whole to pass.

See https://bugs.ruby-lang.org/issues/16131 for details on the
deprecation.

Co-authored-by: Adam Hess <hparker@github.com>